Block quote and CTA renderers
