+++ src/SFA.DAS.FindEmploymentSchemes.Web/wwwroot/css/site.css
@@ -188,7 +188,4 @@ body.js-enabled .app-show-on-mobile-no-js {
 
 .quote-and-attribution {
     font-style: italic;
-    margin: 0 0 20px;
-    padding-left: 20px;
-    border-left: 5px solid #cbd0d3;
 }
